+++ tests/test_visitors/test_ast/test_classes/test_base_classes/test_base_builtin_classes.py
@@ -2,9 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.violations.best_practices import (
-    BuiltinSubclassViolation,
-)
+from wemake_python_styleguide.violations.oop import BuiltinSubclassViolation
 from wemake_python_styleguide.visitors.ast.classes import WrongClassVisitor
 
 class_with_base = """