+++ tests/test_visitors/test_ast/test_classes/test_class_attributes/test_slots_syntax.py
@@ -2,9 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.violations.best_practices import (
-    WrongSlotsViolation,
-)
+from wemake_python_styleguide.violations.oop import WrongSlotsViolation
 from wemake_python_styleguide.visitors.ast.classes import WrongSlotsVisitor
 
 class_body_template = """