+++ tests/test_visitors/test_ast/test_compares/test_nested_ternary.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.violations.best_practices import (
+from wemake_python_styleguide.violations.refactoring import (
     NestedTernaryViolation,
 )
 from wemake_python_styleguide.visitors.ast.compares import (