+++ tests/test_visitors/test_ast/test_exceptions/test_nested_try_blocks.py
@@ -2,9 +2,7 @@
 
 import pytest
 
-from wemake_python_styleguide.violations.best_practices import (
-    NestedTryViolation,
-)
+from wemake_python_styleguide.violations.refactoring import NestedTryViolation
 from wemake_python_styleguide.visitors.ast.exceptions import (
     NestedTryBlocksVisitor,
 )