+++ wemake_python_styleguide/violations/base.py
@@ -41,7 +41,7 @@ Reference
 
 import ast
 import tokenize
-from typing import ClassVar, Optional, Tuple, Union
+from typing import ClassVar, Optional, Set, Tuple, Union
 
 from typing_extensions import final
 
@@ -70,6 +70,7 @@ class BaseViolation(object):
 
     error_template: ClassVar[str]
     code: ClassVar[int]
+    previous_codes: ClassVar[Set[int]]
 
     def __init__(self, node: ErrorNode, text: Optional[str] = None) -> None:
         """