+++ CHANGELOG.md
@@ -9,16 +9,22 @@ We used to have incremental versioning before `0.1.0`.
 ### Features
 
 - **Breaking**: replaces `Z` error code to `WPS` code
+- **Breaking**: creates new violation group `refactoring.py`
+- **Breaking**: creates new violation group `oop.py`
+- **Breaking**: moving a lot of violations
+  from `best_practices` to `refactoring` and `consistency`
 - Adds new `wemake` formatter (using it now by default)
 
 ### Misc
 
+- Adds migration guide
 - Adds `--show-source` to the default recommended configuration
 - Adds better docs about auto-formatters
 - Adds `autopep8` to CI to make sure that `wps` is compatible with it
 - Ensures that `--diff` mode works for `flake8`
 - Renames `Incorrect` to `Wrong` where possible
 - Renames `IncorrectlyNestedTernaryViolation` to `NestedTernaryViolation`
+- Renames `IncorectLoopIterTypeViolation` to `WrongLoopIterTypeViolation`
 
 
 ## 0.10.0 aka The Great Compare